Implemented NewsAPI wrapper class for http request
Added append method to URL class and updated unit tests. All tests pass
