+++ newsfeed/helpers/newsAPI.py
@@ -14,19 +14,21 @@ from helpers.urlBuilder import URL
 
 class NewsAPI:
 
-    API_KEY       = os.environ['NEWS_API_KEY']
     NEWS_API      = "https://newsapi.org/v2/"
     TOP_HEADLINES = "top-headlines?"
+    API_KEY = os.environ['NEWS_API_KEY'] if 'NEWS_API_KEY' in os.environ.keys() else None
 
     ALIASES = { 'keyword' : 'q', 'top' : 'pageSize' }
 
     @classmethod
-    def request(cls, query):
-        url = URL(
-                cls.NEWS_API, cls.TOP_HEADLINES, 
-                query=query, aliases=cls.ALIASES
-                )
+    def __buildURL(cls, query):
+        url = URL(cls.NEWS_API, cls.TOP_HEADLINES, query=query, aliases=cls.ALIASES)
         url.append({'apiKey' : cls.API_KEY})
+        return url
+
+    @classmethod
+    def request(cls, query):
+        url = cls.__buildURL(query)
 
         response = requests.get(url)
